feat(): authorization & authentcation for the review views

+++ backend/review/views.py
@@ -7,6 +7,9 @@ from user.models import User
 from book.models import Book
 from .serializers import ReviewSerializer
 import json
+from rest_framework_simplejwt.authentication import JWTAuthentication
+from rest_framework.decorators import permission_classes
+from rest_framework.permissions import IsAuthenticated
 
 
 @csrf_exempt
@@ -40,22 +43,26 @@ def allreview_noid(request):
 
 
 @csrf_exempt
+@permission_classes([IsAuthenticated])
 def addreview(request, id=None):
     if request.method == 'POST':
+        jwt = JWTAuthentication()
+        try:
+            user, _ = jwt.authenticate(request)
+        except Exception:
+            return JsonResponse({'error': 'You need to Login'}, status=401)
         if request.content_type == 'application/json':
-            mandatory_keys = ['id', 'title', 'media', 'content', 'rating']
+            mandatory_keys = ['title', 'media', 'content', 'rating']
             missing_keys = []
             if id is not None:
                 temp = json.loads(request.body)
                 missing_keys = [key for key in mandatory_keys if key not in temp]
                 if missing_keys:
                     return JsonResponse({'error': f'Missing Key(s): {", ".join(missing_keys)}'}, status=400)
-                owner_id = temp.get('id')
-                reviewer = User.objects.get(pk=owner_id)
                 book = Book.objects.get(pk=id)
                 book_name = book.name
                 new_review = Review.objects.create(
-                    owner=reviewer,
+                    owner=user,
                     title=temp['title'],
                     book=book,
                     media=temp['media'],
@@ -65,6 +72,7 @@ def addreview(request, id=None):
                 serialize = ReviewSerializer(new_review)
                 response_data = serialize.data
                 response_data['book_title'] = book_name
+                response_data['reviewer'] = user.username
                 return JsonResponse(response_data, safe=False, status=201)    
             else:
                 return JsonResponse({'error': 'No Book to review'})
@@ -75,14 +83,23 @@ def addreview(request, id=None):
 
 
 @csrf_exempt
+@permission_classes([IsAuthenticated])
 def deletereview(request, id=0):
     if request.method == 'DELETE':
+        jwt = JWTAuthentication()
+        try:
+            user, _ = jwt.authenticate(request)
+        except Exception:
+            return JsonResponse({'error': 'You need to Login'}, status=401)
         try:
             review_id = int(id)
             try:
                 review = Review.objects.get(pk=review_id)
-                review.delete()
-                return JsonResponse({'message': 'Review Deleted!'}, status=200)
+                if user == review.owner:
+                    review.delete()
+                    return JsonResponse({'message': 'Review Deleted!'}, status=200)
+                else:
+                    return JsonResponse({'error': 'Unauthorized'}, status=403)
             except ObjectDoesNotExist:
                 return JsonResponse({'error': 'No such Object'}, status=400)
         except ValueError:
